% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV.R
\name{EV}
\alias{EV}
\title{Expected Values}
\usage{
EV(c, n, pn, lambda, type = "")
}
\arguments{
\item{c}{number of servers}

\item{n}{number of customers}

\item{pn}{vector of probabilities of having n customers in the system}

\item{lambda}{arrival rate}

\item{type}{character, type in E_Nq, E_N, E_Wq, E_W for the desired expected value}
}
\value{
Expected Values either of E_Nq, E_N, E_Wq, E_W
}
\description{
Returns the expected value depending on the entry for type
}
\examples{
EV(c=1, n=4, pn=c(0.410, 0.274, 0.167, 0.097, 0.052), lambda=20, type="E_Nq")
}
