% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PN.R
\name{PN}
\alias{PN}
\title{Pn Function}
\usage{
PN(n, lambdas, mus)
}
\arguments{
\item{n}{number of effective customers/objects}

\item{lambdas}{vector of dependent arrival rates !IMPORTANT! MUST HAVE LENGTH N!}

\item{mus}{vector of dependent service rates !IMPORTANT! MUST HAVE LENGTH N!}
}
\value{
Probability
}
\description{
Returns the probabilities of having n customers in the system
}
\examples{
PN(n=4, lambdas=rep(20, 4), mus=c(30, 32.72, 34.48, 37.39))
}
